refactor: `ExampleServer` internal module
